LayoutTests/inspector/debugger/script-snippet-model-expected.txt

Tests script snippet model.

Debugger was enabled.

Running: testWorkspaceReset
Snippet created.
Resetting snippet script mapping.
Snippet uiSourceCode was recreated.

Running: testEvaluate
Last evaluation source url for snippet: snippets:///1_1
Snippet script added, sourceURL = snippets:///1_1
Snippet execution result: undefined
Last evaluation source url for snippet: snippets:///2_2
Snippet script added, sourceURL = snippets:///2_2
Snippet execution result: function doesNothing() {
    var  i = 2+2;
}
Last evaluation source url for snippet: snippets:///1_3
Snippet script added, sourceURL = snippets:///1_3
Snippet execution result: undefined
Debugger was disabled.
